Terminado ajustes eliminar autor y editorial

+++ src/main/java/egg/ejercicio01/libreria/servicios/AutorServicio.java
@@ -9,13 +9,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import egg.ejercicio01.libreria.entidades.Autor;
+import egg.ejercicio01.libreria.entidades.Libro;
 import egg.ejercicio01.libreria.errores.ErrorServicio;
 import egg.ejercicio01.libreria.repositorios.AutorRepositorio;
+import egg.ejercicio01.libreria.repositorios.LibroRepositorio;
 
 @Service
 public class AutorServicio {
     @Autowired // Spring inicializa automaticamente
     private AutorRepositorio autorRepositorio;
+
+    @Autowired
+    private LibroRepositorio libroRepositorio;
 
     // esta anotacion hace commit automatico si esta todo en orden, caso contrario
     @Transactional // hace roollback
@@ -53,18 +58,34 @@ public class AutorServicio {
 
     @Transactional
     public void deleteAutor(String id) throws ErrorServicio {
-        Optional<Autor> respuesta = autorRepositorio.findById(id);
-        if (respuesta.isPresent()) {
-            autorRepositorio.deleteById(id);
-        }else {
-            throw new ErrorServicio("No existe el autor con id: " + id);
+        String librosAsociados = "";
+        int count = 0;
+        for (Libro l : libroRepositorio.findAll()) {
+            if (l.getAutor().getId().equals(id)) {
+                count += 1;
+                librosAsociados += l.getTitulo() + ", ";
+            }
         }
+        
+        if (count > 0) {
+            librosAsociados = librosAsociados.substring(0, librosAsociados.length() - 2) + ".";
+            throw new ErrorServicio("No se puede eliminar el autor porque tiene libros asociados: " + librosAsociados);
+        } else {
+
+            Optional<Autor> respuesta = autorRepositorio.findById(id);
+            if (respuesta.isPresent()) {
+                autorRepositorio.deleteById(id);
+            } else {
+                throw new ErrorServicio("No se se encontro un autor con el id: " + id);
+            }
+        }
+
     }
 
     @Transactional
     public void disableAutor(String id) {
         Optional<Autor> respuesta = autorRepositorio.findById(id);
-        if (respuesta.isPresent()) { 
+        if (respuesta.isPresent()) {
             Autor autor = respuesta.get();
             autor.setAlta(false);
             autorRepositorio.save(autor); // este save no crea id, actualiza
